refactor: merge data loader logic in one file
